prepared css properties for layout engine

+++ Sparky-core/src/sp/css/CSSManager.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <sp/sp.h>
 #include <SparkyCSS/SparkyCSS.h>
 
 namespace sp { namespace css {
@@ -11,8 +12,8 @@ namespace sp { namespace css {
 		CSSManager();
 		~CSSManager();
 
-		inline float GetWidth() const { return Application::GetApplication().GetWindowWidth(); }
-		inline float GetHeight() const { return Application::GetApplication().GetWindowHeight(); }
+		inline uint GetWidth() const { return Application::GetApplication().GetWindowWidth(); }
+		inline uint GetHeight() const { return Application::GetApplication().GetWindowHeight(); }
 
 		void EvalCSS(const String& css);
 		CSSValue* GetValue(const UIElementCSSInfo& cssinfo, CSSKey key);
